Police Computer
To sum up the Police computer data base program and its main aim was to create the program that
could help to create useful database for finding lost documents.The project was implemented in C++
using object-oriented programming concepts, such as inheritance, polymorphism. Database was
represented by a vector of pointers to Document objects, which was a base class for three derived
classes: Passport, IdentityCard, and DrivingLicense. The program also provided a menu-driven interface
for the user to perform various operations on the database, such as searching, updating, adding, and
deleting documents. The program was tested in many cases and showed satisfactory results